created home btn and added new folder for routes
